% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{airline}
\alias{airline}
\title{Airline passenger data}
\format{
A tibble with 1,979 rows and 2 columns:
\describe{
\item{year}{Two digit year.}
\item{month}{Numerical month.}
\item{passengers}{Number of passengers.}
\item{time}{Ordered of observations.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
airline
}
\description{
Daily data on the number of airline passengers.
}
\keyword{datasets}
