% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{browser_data}
\alias{browser_data}
\title{Online-spending activity and demographics for a sample of households.}
\format{
A data frame of 10,000 rows representing hourly measurements and 15 variables
\describe{
\item{id}{Household id number.}
\item{anychildren}{Indicator variable for any children in the household.}
\item{broadband}{Relative humidity.}
\item{hispanic}{Indicator variable for Hispanic.}
\item{race}{Household race classification.}
\item{region}{Household region.}
\item{spend}{Online spending ($).}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
browser_data
}
\description{
Online-spending activity and demographics for a sample of households.
}
\keyword{datasets}
