% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{credit}
\alias{credit}
\title{German credit default data from Professor Hoffman.}
\format{
A tibble with 1,000 rows and 9 columns:
\describe{
\item{default}{Binary variable indicating whether the client defaulted or not}
\item{duration}{Duration of loan in months}
\item{amount}{Loan amount}
\item{installment}{Number of installments}
\item{age}{Client age in years}
\item{history}{Credit history}
\item{purpose}{Loan purpose}
\item{foreign}{German or foreign}
\item{rent}{Boolean for renter or not}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
credit
}
\description{
Data set with default and client characteristics.
Can be used to practice creating a statistical algorithm for banks on whether to extend credit card loans or not.
}
\keyword{datasets}
