% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dja}
\alias{dja}
\title{Dow Jones Industrial Average data}
\format{
A tibble with 1,979 rows and 2 columns:
\describe{
\item{dja}{Index value.}
\item{day}{Ordered day of observation.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
dja
}
\description{
Daily data on the Dow Jones Industrial Average, which is a stock market index of 30 U.S. prominent companies.
}
\keyword{datasets}
