% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fnr}
\alias{fnr}
\title{false negative rate: pr(p = 0 | y = 1)}
\usage{
fnr(.p, .model)
}
\arguments{
\item{.p}{probability}

\item{.model}{a model}
}
\description{
false negative rate: pr(p = 0 | y = 1)
}
