% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr}
\alias{fpr}
\title{false positive rate: pr(p = 1 | y = 0)}
\usage{
fpr(.p, .model)
}
\arguments{
\item{.p}{probability}

\item{.model}{a model}
}
\description{
false positive rate: pr(p = 1 | y = 0)
}
