% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gas}
\alias{gas}
\title{Octane and near-infrared (NIR) spectrometry measurements}
\format{
A tibble with 60 rows and 402 columns:
\describe{
\item{octane}{Octane rating.}
\item{nir_}{Near-infrared (NIR) spectrometry measures.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
gas
}
\description{
Octane and near-infrared (NIR) spectrometry measurements
}
\keyword{datasets}
