% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP and imports data by country in 2022.}
\format{
A tibble with 174 rows and 4 columns:
\describe{
\item{country}{Country name.}
\item{year}{Year.}
\item{gdp}{Gross domestic product, millions $ (2022).}
\item{imports}{Value of imports, millions $ (2022).}
}
}
\source{
Wikipedia.
}
\usage{
gdp
}
\description{
GDP and imports data by country in 2022.
}
\keyword{datasets}
