% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oj}
\alias{oj}
\title{Orange juice sales data collected in the 1990s from Dominick's, which was a Chicago-area chain store.}
\format{
A tibble with 28,947 rows and 4 columns:
\describe{
\item{sales}{Sales volume ($).}
\item{price}{Price ($).}
\item{brand}{Orange juice brand.}
\item{feat}{Indicator for whether brand was advertised that week or not.}
}
}
\source{
Kilts Center at the University of Chicago's Booth School of Business.
}
\usage{
oj
}
\description{
Orange juice sales data collected in the 1990s from Dominick's, which was a Chicago-area chain store.
}
\keyword{datasets}
