% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SC}
\alias{SC}
\title{Quality-control data from a semiconductor manufacturing process.}
\format{
A tibble with 1,477 rows and 201 columns:
\describe{
\item{FAIL}{Indicator variable for whether the observed chip failed or not.}
\item{SIG_}{Measurement from diagnostic sensor _ in the production process.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
SC
}
\description{
Quality-control data from a semiconductor manufacturing process.
}
\keyword{datasets}
