% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp}
\alias{sp}
\title{S&P500 monthly returns}
\format{
A tibble with 84 rows and 2 columns:
\describe{
\item{month}{Month.}
\item{r_sp}{Monthly return.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
sp
}
\description{
S&P500 monthly returns
}
\keyword{datasets}
