% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split}
\alias{split}
\title{function to split data frame into train and test data frames}
\usage{
split(df, p = 0.8)
}
\arguments{
\item{.p}{probability}

\item{.model}{a model}
}
\description{
function to split data frame into train and test data frames
}
