% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Stock data}
\format{
A tibble with 42,355 rows and 3 columns:
\describe{
\item{ticker}{Stock ticker.}
\item{month}{Month.}
\item{r}{Monthly return.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
stocks
}
\description{
Monthly returns
}
\keyword{datasets}
