% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_lasso.cv.gamlr}
\alias{tidy_lasso.cv.gamlr}
\title{tidy as gamlr::gamlr() object}
\usage{
\method{tidy_lasso}{cv.gamlr}(lasso)
}
\arguments{
\item{lasso}{a model of class cv.gamlr}
}
\description{
This model is not included in broom::tidy() so I made a similar function
}
\keyword{cv.gamlr}
\keyword{lasso}
\keyword{tidy}
