% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_lasso.gamlr}
\alias{tidy_lasso.gamlr}
\title{tidy as gamlr::gamlr.cv() object}
\usage{
\method{tidy_lasso}{gamlr}(lasso)
}
\arguments{
\item{lasso}{a model of class gamlr}
}
\description{
This model is not included in broom::tidy() so I made a similar function
}
\keyword{gamlr}
\keyword{lasso}
\keyword{tidy}
