% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_lasso}
\alias{tidy_lasso}
\title{tidy some models}
\usage{
tidy_lasso(lasso)
}
\arguments{
\item{lasso}{a model from gamlr or cv.gamlr}
}
\description{
Some additional models not included in broom::tidy()
}
\keyword{lasso}
\keyword{tidy}
