% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytaddy-package.R
\docType{package}
\name{tidytaddy-package}
\alias{tidytaddy}
\alias{tidytaddy-package}
\title{tidytaddy: Data sets used in Business Data Science by Matt Taddy}
\description{
This puts all the data sets used in Business Data Science by Matt Taddy into a single package. I use this for a student seminar where we go through the book and I wanted all the data sets to be easily accessible for the students.
}
\author{
\strong{Maintainer}: Andrew Griffen \email{andrew.griffen@gmail.com}

}
\keyword{internal}
