% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tnr}
\alias{tnr}
\title{true negative rate: pr(p = 0 | y = 0)}
\usage{
tnr(.p, .model)
}
\arguments{
\item{.p}{probability}

\item{.model}{a model}
}
\description{
true negative rate: pr(p = 0 | y = 0)
}
