% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tpr}
\alias{tpr}
\title{true positive rate:  pr(p = 1 | y = 1)}
\usage{
tpr(.p, .model)
}
\arguments{
\item{.p}{probability}

\item{.model}{a model}
}
\description{
true positive rate:  pr(p = 1 | y = 1)
}
