% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trucks}
\alias{trucks}
\title{Data on pickup trucks.}
\format{
A tibble with 46 rows and 4 columns:
\describe{
\item{year}{Vintage.}
\item{miles}{Odometer miles.}
\item{price}{Price.}
\item{make}{Make.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
trucks
}
\description{
Data on pickup trucks.
}
\keyword{datasets}
