% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{abortion}
\alias{abortion}
\title{Abortion and crime data.}
\format{
A tibble with 1,734 rows and 17 columns:
\describe{
\item{murder}{De-trended log murder rate.}
\item{violent}{De-trended log violent crime rate.}
\item{property}{De-trended log property crime rate.}
\item{state}{State.}
\item{prisoners}{log of lagged prisoners per capita.}
\item{police}{log of lagged police per capita.}
\item{unemployment}{the unemployment rate.}
\item{income}{per-capita income.}
\item{poverty}{the poverty rate.}
\item{AFDC}{AFDC generosity at year t-15.}
\item{gun}{Indicator for concealed weapons law.}
\item{beer}{Beer consumption per capita.}
}
}
\source{
Donohue and Levitt (2001).
}
\usage{
abortion
}
\description{
Data set from "The Impact Of Legalized Abortion On Crime" by John J. Donohue III And Steven D. Levitt.
}
\keyword{datasets}
