% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{email}
\alias{email}
\title{Data from George's inbox.}
\format{
A tibble with 4,601 rows and 58 columns:
\describe{
\item{spam}{Indicator variable for whether the email was classified as spam or not.}
\item{word_}{Indicator variable for the presence of the word_.}
}
}
\source{
Business Data Science by Matt Taddy.
}
\usage{
email
}
\description{
Data set created from emails in a guy named George's inbox.
Each email has been classified as spam or not and indicator variables (or one-hot encodings)
have been created for the most common words in the emails. Commonly used data to practice
creating a spam filter.
}
\keyword{datasets}
